+++ flutter_searchbox_ui/lib/flutter_searchbox_ui.dart
@@ -2,3 +2,4 @@ library flutter_searchbox_ui;
 
 export 'src/components/range_input.dart';
 export 'src/components/reactive_google_map.dart';
+export 'src/components/selected_filters.dart';